slight refactor + add docs
