+++ src/requests/cover.rs
@@ -1,8 +1,15 @@
+//! Utilities for scanlation group specific server part of the application.
+//!
+//! The meaning of most of the confusing structs here can be found at <https://api.mangadex.org/docs/3-enumerations/#manga-links-data>
+//!
+//! All queries encountered here can be constructed with the builder syntax from the [bon] crate
+
 use serde::{Deserialize, Serialize};
 
 use super::query_utils::Relationship;
 use super::{EntityType, Locale};
 
+/// Used for serialization/deserialization
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct CoverArtAttributes {
@@ -15,6 +22,7 @@ pub struct CoverArtAttributes {
     pub updated_at: String,
 }
 
+/// Main structure used for representing the response containg cover info
 #[derive(Serialize, Deserialize, Debug)]
 pub struct CoverArt {
     pub id: String,
