+++ src/requests/scanlation_group.rs
@@ -1,8 +1,15 @@
+//! Utilities for scanlation group specific server part of the application.
+//!
+//! The meaning of most of the confusing structs here can be found at <https://api.mangadex.org/docs/3-enumerations/#manga-links-data>
+//!
+//! All queries encountered here can be constructed with the builder syntax from the [bon] crate
+
 use serde::{Deserialize, Serialize};
 
 use super::query_utils::{LocalizedString, Relationship};
 use super::{Entity, EntityType, Locale};
 
+/// Used for serialization/deserialization
 #[derive(Serialize, Deserialize, Clone, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct ScanlationGroupAttributes {
@@ -27,6 +34,7 @@ pub struct ScanlationGroupAttributes {
     pub updated_at: String,
 }
 
+/// Main structure used for representing the response containg scanlation group info
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct ScanlationGroup {
     pub id: String,
